fixing catching errors

+++ app/main.py
@@ -48,11 +48,15 @@ class ProductsTable:
         return self.records
 
     def search(self, id):
+        if (isinstance(id, int) == False):
+            return jsonify({"msg": f"Bad Request 400: the provided id is not integer, or it contains illegal form of characters", "statCode": 400})
         self.cur.execute(f"SELECT * FROM products WHERE id = '{id}'")
         self.record = self.cur.fetchone()
         return self.record
 
     def insert(self, id, title, price, img):
+        if (isinstance(id, int) == False or isinstance(price, int) == False):
+            return jsonify({"msg": f"Bad Request 400: the provided id or price is not integer, or it contains illegal form of characters", "statCode": 400})
         if (id == "" or price == "" or title == ""):
             raise Exception("One of the entries is empty")
         self.cur.execute(f"""
@@ -61,6 +65,8 @@ class ProductsTable:
         self.conn.commit()
 
     def update(self, id, title, price, img):
+        if (isinstance(id, int) == False or isinstance(price, int) == False):
+            return jsonify({"msg": f"Bad Request 400: the provided id or price is not integer, or it contains illegal form of characters", "statCode": 400})
         self.cur.execute(
             f"UPDATE products SET title = '{title}' WHERE id = '{id}'")
         self.cur.execute(
@@ -70,6 +76,8 @@ class ProductsTable:
         self.conn.commit()
 
     def delete(self, id):
+        if (isinstance(id, int) == False):
+            return jsonify({"msg": f"Bad Request 400: the provided id or price is not integer, or it contains illegal form of characters", "statCode": 400})
         if (id == None):
             raise Exception("You have to select an id to delete its values")
         self.cur.execute(f"DELETE FROM products WHERE id = '{id}'")
@@ -94,11 +102,15 @@ class PromocodesTable:
         return self.records
 
     def search(self, id):
+        if (isinstance(id, int) == False):
+            return jsonify({"msg": f"Bad Request 400: the provided id is not integer, or it contains illegal form of characters", "statCode": 400})
         self.cur.execute(f"SELECT * FROM promocodes WHERE id = '{id}'")
         self.record = self.cur.fetchone()
         return self.record
 
     def insert(self, id, code, amount):
+        if (isinstance(id, int) == False or isinstance(amount, float) == False):
+            return jsonify({"msg": f"Bad Request 400: the provided id is not integer or amount is not float, or it contains illegal form of characters", "statCode": 400})
         if (id == "" or code == "" or amount == ""):
             raise Exception("One of the entries is empty")
         self.cur.execute(f"""
@@ -107,6 +119,8 @@ class PromocodesTable:
         self.conn.commit()
 
     def update(self, id, code, amount):
+        if (isinstance(id, int) == False or isinstance(amount, float) == False):
+            return jsonify({"msg": f"Bad Request 400: the provided id is not integer or amount is not float, or it contains illegal form of characters", "statCode": 400})
         self.cur.execute(
             f"UPDATE promocodes SET code = '{code}' WHERE id = '{id}'")
         self.cur.execute(
@@ -114,6 +128,8 @@ class PromocodesTable:
         self.conn.commit()
 
     def delete(self, id):
+        if (isinstance(id, int) == False):
+            return jsonify({"msg": f"Bad Request 400: the provided id is not integer, or it contains illegal form of characters", "statCode": 400})
         if (id == None):
             raise Exception("You have to select an id to delete its values")
         self.cur.execute(f"DELETE FROM promocodes WHERE id = '{id}'")
@@ -214,6 +230,8 @@ class StoreNumTable:
         return self.record
 
     def insert(self, storeNum):
+        if (isinstance(storeNum, int) == False):
+            return jsonify({"msg": f"Bad Request 400: the provided storeNum is not integer, or it contains illegal form of characters", "statCode": 400})
         if (storeNum == ""):
             raise Exception("One of the entries is empty")
         self.cur.execute(f"""
@@ -223,11 +241,15 @@ class StoreNumTable:
         self.conn.commit()
 
     def update(self, storeNum):
+        if (isinstance(storeNum, int) == False):
+            return jsonify({"msg": f"Bad Request 400: the provided storeNum is not integer, or it contains illegal form of characters", "statCode": 400})
         self.cur.execute(
             f"UPDATE storeNums SET storeNum = '{storeNum}'")
         self.conn.commit()
 
     def delete(self, storeNum):
+        if (isinstance(storeNum, int) == False):
+            return jsonify({"msg": f"Bad Request 400: the provided storeNum is not integer, or it contains illegal form of characters", "statCode": 400})
         if (storeNum == None):
             raise Exception("You have to select an id to delete its values")
         self.cur.execute(
